+++ package.json
@@ -22,6 +22,7 @@
     "date-fns": "^2.30.0",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "fasttext": "^1.0.0",
     "ix": "^5.0.0",
     "jws": "^4.0.0",
     "kysely": "^0.25.0",
